Refactor: add a new directory called validations and move all validations to her root, move tokengenerator to utils and fix all imports
