+++ back-end/src/Users/Users.models.ts
@@ -9,11 +9,7 @@ const getUsers = async () => {
 
 // id it's a cuid generated by Prisma
 const getUserById = async (id: string) => {
-  const user = await prisma.users.findUnique({
-    where: {
-      id
-    }
-  })
+  const user = await prisma.users.findUnique({ where: { id } });
   if(!user) {
     throw new CustomError("User not found", BAD_REQUEST.statusCode);
   }
