+++ back-end/src/Validations/handleRegister/isEmailValid.ts
@@ -1,4 +1,4 @@
-import { EMAIL_REGEX } from "../constants";
+import { EMAIL_REGEX } from "../../utils/constants";
 
 const isEmailValid = (email: string) => {
   return EMAIL_REGEX.test(email);
