+++ back-end/src/Validations/handleRegister/isNameValid.ts
@@ -1,4 +1,4 @@
-import { NAME_REGEX } from "../constants";
+import { NAME_REGEX } from "../../utils/constants";
 
 const isNameValid = (name: string) => {
   return NAME_REGEX.test(name);
