+++ back-end/src/Validations/handleRegister/passwordEncrypter.ts
@@ -1,5 +1,5 @@
 import bcrypt from "bcrypt";
-import { SALT_ROUNDS } from "../constants";
+import { SALT_ROUNDS } from "../../utils/constants";
 
 const passwordEncrypter = async (password: string) => {
   const hashedPassword = await bcrypt.hash(password, SALT_ROUNDS);
